Remove UpdateByIdTask indirection
This inlines the logic into the UpdateById plan. The separate task is a
leftover from when we still had a Task interface for the execution
logic.
